requirements.txt
geopandas==0.13.2
geopy==2.3.0
matplotlib==3.7.2
numpy==1.24.3
osmium==3.6.0
pandas==2.0.1
Pillow==10.0.0
pyproj==3.6.0
Requests==2.31.0
scipy==1.11.2
Shapely==1.8.1
rasterio==1.3.8